backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
torch==2.0.1
torchvision==0.15.2
transformers==4.33.2
Pillow==10.4.0
numpy==1.24.3
gunicorn==21.2.0
requests==2.31.0
pytesseract==0.3.10
opencv-python==4.8.1.78
easyocr==1.7.0
openai>=1.0.0
httpx>=0.25.0
timm==1.0.19
